% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arthritis}
\alias{Arthritis}
\title{Data for example}
\format{
A data.frame with 84 obs. and 5 var.
}
\usage{
Arthritis
}
\description{
This data are imported from the package vcd, which depends on a
package (i.e., mvtnorm) which require R >= 3.5. For this reason
we do not import the package vcd, but only this dataset. Which is
used only in the examples and tests.
}
\keyword{datasets}
