% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_bot_options.R
\name{check_for_bot_options}
\alias{check_for_bot_options}
\title{Check if a bot is set up}
\usage{
check_for_bot_options(chat_id, bot)
}
\arguments{
\item{chat_id}{a chat id}

\item{bot}{a bot}
}
\value{
invisible lgl indicating if check pass (TRUE) or not (FALSE)
}
\description{
\code{\link{check_for_bot_options}} check if a telegram bot and
corresponding chat, exist.
}
