% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_pkg}
\alias{check_pkg}
\title{Check basic installed packages}
\usage{
check_pkg(interested = ubesp_pkg, dependencies = TRUE)
}
\arguments{
\item{interested}{[chr] packages' name}

\item{dependencies}{do you want to install the dependencies?}
}
\value{
invisible character vector with missing package
}
\description{
Check basic installed packages
}
