% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depigner-package.R
\docType{package}
\name{depigner-package}
\alias{depigner}
\alias{depigner-package}
\title{Utilities for _pigna_s}
\description{
This package collect usefull function frequently used at the Unit of
Biostatistics, Epidemiology and Public Health of the Department of Cardiac,
Thoracic and Vascular sciences at the university of Padova.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corradolanera.github.io/depigner/}
  \item \url{https://github.com/CorradoLanera/depigner}
  \item Report bugs at \url{https://github.com/CorradoLanera/depigner/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} [copyright holder]

}
\keyword{internal}
