% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depigner.R
\docType{package}
\name{depigner}
\alias{depigner}
\alias{depigner-package}
\title{Utilities for _pigna_s}
\description{
This package collect usefull function frequently used at the Unit of
Biostatistics, Epidemiology and Public Health of the Department of Cardiac,
Thoracic and Vascular sciences at the university of Padova.
}
