% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{gdp}
\alias{gdp}
\title{GDP}
\usage{
gdp(time = 4L, freq = 0.05)
}
\arguments{
\item{time}{[num] number of seconds needed to relax yourself
(default = 4)}

\item{freq}{[num] rotating frequency (default = 0.05/s)}
}
\value{
invisible()
}
\description{
A wrapper to relax
}
\examples{
if (interactive()) {
  gdp()

  gdp(5)

  required_seconds_of_relax <- 10 # AKA: "conta fino a dieci!"
  rate_of_anxiety <- 1/1000       # AKA: "mi girano a mille!"
  gdp(required_seconds_of_relax, rate_of_anxiety)
}
}
