% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_from.R
\name{imported_from}
\alias{imported_from}
\title{Packages imported by a package}
\usage{
imported_from(x, include_self = FALSE)
}
\arguments{
\item{x}{(chr) the package you would like to see its imports}

\item{include_self}{(lgl, FALSE) do you like to include `x` itself in
the output?}
}
\value{
(chr) vector of packages imported by `x`
}
\description{
Packages imported by a package
}
\examples{
imported_from("depigner")
}
