% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkg_set.R
\name{install_pkg_set}
\alias{install_pkg_set}
\title{Check basic installed packages}
\usage{
install_pkg_set(set = pkg_all, dependencies = TRUE)
}
\arguments{
\item{set}{(chr) packages' names}

\item{dependencies}{do you want to install the dependencies?}
}
\value{
invisible character vector of the subset of `interested`
  which was not already present, and installed.
}
\description{
This function is used to politely ask the user to install requiring
packages. It is intended to be used in interactive sessions only.
}
\details{
You can pass arbitrarily sets of packages; on the other hands, you
can use some sets already prepared and included into `{depigner}`
(see `?pkg_sets`).
}
\note{
By default this function install all the packages listed in
  `pkg_all`, which is the union of all the sets of packages listed
  in `?pkg_sets`.
}
\examples{
\dontrun{
  install_pkg_set() # to install all the `?pkg_all`
}
}
\seealso{
please_install, pkg_sets
}
