% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_telegram_bot.R
\name{prepare_telegram_bot}
\alias{prepare_telegram_bot}
\title{Set up a Telegram bot}
\usage{
prepare_telegram_bot(bot_name = NULL, chat_id = NULL)
}
\arguments{
\item{bot_name}{(chr, NULL) This argument should be left NULL. If
NULL, the function bring the bot name from the environmental
variable "R_telegram_bot_name". You can pass your bot's name
here as a character string too.}

\item{chat_id}{(chr, NULL) This argument could be left NULL. If NULL,
the funciton bring the first chat id from the ones used by the
bot.}
}
\value{
invisibly a list with the bot name and the chat_id used.
}
\description{
Produce a side effect to assign to the Global Environment two hidden
object for the bot and its chat_id
}
\examples{
\dontrun{
    library(depigner)
    prepare_telegram_bot()
}
}
