% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_summary.R
\name{summary_cox}
\alias{summary_cox}
\title{summary_cox}
\usage{
summary_cox(cox_model)
}
\arguments{
\item{cox_model}{A `code{cph}` object as produced by \code{\link[rms]{cph}}.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Convert the output of the \code{\link[rms]{summary.rms}} into a data
frame, reporting only the Hazard Ratio with the .95 CI and the incremental
step (for continuous variables) reference (for categorical variables) for
which the Hazard is refered to (i.e. without $\beta$s, Low, High, S.E. and
Type).
}
\note{
This is a non-exported function, use this at your own risk.
Probably it will change in the near future.
}
