% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_summary.R
\name{tidy_summary}
\alias{tidy_summary}
\alias{tidy_summary.summary.formula.reverse}
\title{tidy_summary}
\usage{
tidy_summary(x, ...)

\method{tidy_summary}{summary.formula.reverse}(x, ...)
}
\arguments{
\item{x}{an object used to select a method, output of some summary
by \code{\link{Hmisc}}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Converts a \code{summary()} object produced by \code{Hmisc} or
by \code{rms} packages to a tidy data frame ready to be
\code{\link[pander]{pander}}ed (e.g. printed on a word document after
\code{\link[knitr]{knit}} the source.
}
\section{Methods (by class)}{
\itemize{
\item \code{summary.formula.reverse}: tidy a summary reverse output from the
\code{\link[Hmisc]{summary.formula}} called with
\code{method = "reverse"}.
}}

\note{
to see the options you can pass to \code{...} for a custom
print, see the print section in \code{\link[Hmisc]{summary.formula}}.
}
\examples{
library(Hmisc)
my_summary <- summary(Species ~., data = iris, method = "reverse")
tidy_summary(my_summary)
}
