% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubesp_pkg}
\alias{ubesp_pkg}
\title{Package used at UBESP.}
\format{An object of class \code{character} of length 42.}
\usage{
ubesp_pkg
}
\description{
Character vector of the main packages uses at UBESP
}
\keyword{datasets}
