% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ui.R
\name{use_ui}
\alias{use_ui}
\title{Use usethis'ui(s) in your package}
\usage{
use_ui()
}
\description{
Does setup necessary to use the usethis' user interfaces in your package.
This function requires the use roxygen.
* Check that the active package uses roxygen2
* Adds usethis package to "Imports" in `DESCRIPTION`
* Imports in your namespace:
  - block styles: \code{\link{ui_line}}, \code{\link{ui_todo}}
      \code{\link{ui_done}}, \code{\link{ui_todo}}, \code{\link{ui_oops}}
      \code{\link{ui_info}}, \code{\link{ui_code_block}}
  - conditions: \code{\link{ui_stop}}, \code{\link{ui_warn}}
  - questions: \code{\link{ui_yeah}}, \code{\link{ui_nope}}
  - inline styles: \code{\link{ui_field}}, \code{\link{ui_value}}
      \code{\link{ui_path}}, \code{\link{ui_code}}) user interfaces
}
\details{
Attribution: most of the source content of this function is
  taken and/or adapted from the corresponding unexported function in
  the `usethis` package.
}
\examples{
\dontrun{
  # while setup of a package
  use_ui()
}
}
