% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_to_telegram.R
\name{errors_to_telegram}
\alias{errors_to_telegram}
\title{Title}
\usage{
errors_to_telegram(chat_name = Sys.getenv("R_telegram_error_chat_name"),
  bot_name = getOption("depigner.bot_name"))
}
\arguments{
\item{chat_name}{a}

\item{bot_name}{a}
}
\value{
a
}
\description{
Title
}
\examples{
\dontrun{
    library(depigner)
    errors_to_telegram()
}
}
