% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{gdp}
\alias{gdp}
\title{GDP}
\usage{
gdp(time = 4, freq = 0.05)
}
\arguments{
\item{time}{[num] number of seconds needed to relax yourself
(default = 4)}

\item{freq}{[num] rotating frequency (default = 0.05/s)}
}
\value{
nothing
}
\description{
A wrapper to relax
}
\examples{
gdp()
}
