% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Htypes.R
\name{Htype}
\alias{Htype}
\alias{Htypes}
\alias{Htypes.describe}
\alias{Htypes.default}
\alias{is_Hcat}
\alias{is_Hcon}
\title{Type's checks accordingly to [Hmisc] package}
\usage{
Htype(x, n.unique = 10)

Htypes(x, n.unique = 10)

\method{Htypes}{describe}(x, n.unique = 10)

\method{Htypes}{default}(x, n.unique = 10)

is_Hcat(x)

is_Hcon(x, n.unique = 10)
}
\arguments{
\item{x}{an instance of the class [describe][Hmisc::describe], in the
cases of "singoular" functions (ie `is_*()` or `Htype()`) it must
be a single-variable [describe][Hmisc::describe] object.}

\item{n.unique}{(int, 10) the minimum number of distinct values a
numeric variable must have before plot.describe uses it in a
continuous variable plot.}
}
\value{
(chr) `Htype` returns one of "cat" (if `x` will be considered
  categorical), "con" (if `x` will be considered continuous), "none"
  (if `x` will be not considered categorical nor continuous, and
  hence it will be not plotted), or "both" (with a warning, if the
  variable will be considered both categorical and continuous. This
  would possibly nevere happen).

(chr) character vector of the types identified by [Htype] for
  every variable represented in (each element of) `x`.

(lgl) `is_Hcat` returns TRUE if x will be considered
  categorical.

(lgl) `is_Hcon` returns TRUE if x will be considered
  continuous.
}
\description{
These functions decide and report if a single variable represented by
a single instance of an `Hmisc`'s [describe][Hmisc::describe] object
will considered a categorical variable or a continuous one.
}
\details{
A "single" object of `Hmisc`'s [describe][Hmisc::describe]
  class represents a variable. When you plot and object of class
  [describe][Hmisc::describe] the plot function decide if it is a
  continuous variable or a categorical one to plot it in the
  correspond plot. It is also possible that the variable is not
  considered in none of that categories, in which case it will not be
  plotted at all.

  These functions have been produced/deduced from reading the
  source code of `Hmisc`'s [plot][Hmisc::describe]. In particular,
  from the definition of the (two distinct) functions `f` defined
  within it (one for categorical variables and the other for
  continuous variables). Both lead to a possible execution of
  `warning("no categorical variable found")` or `warning("no
  continuous variable found")`. I tried to keep the same
  names/code/logic that I found there.
}
\section{Functions}{
\itemize{
\item \code{Htypes}: Report types for multi-variables objects

\item \code{is_Hcat}: Check if a single-instance of a
[describe][Hmisc::describe] object is categorical.

\item \code{is_Hcon}: Check if a single-instance of a
[describe][Hmisc::describe] object is continuous.
}}

\examples{
library(Hmisc)
desc <- describe(mtcars)

Htype(desc[["vs"]])  # "cat"
Htype(desc[["mpg"]]) # "con"
Htype(desc[["cyl"]]) # "none"

Htypes(desc)  # c(
              #   mpg = "con", cyl = "none", disp = "con",
              #   hp = "con", drat = "con", wt = "con", qsec = "con",
              #   vs = "cat", am = "cat", gear = "none",
              #   carb = "none"
              # )
Htypes(mtcars)  # Htypes(desc)
Htypes(letters) # "none"
is_Hcat(desc[["vs"]])  # TRUE
is_Hcat(desc[["mpg"]]) # FALSE
is_Hcon(desc[["vs"]])  # FALSE
is_Hcon(desc[["mpg"]]) # TRUE
}
\seealso{
[describe][Hmisc::describe],

[is_Hdesc], [is_single_Hdesc]

Gist with test and usage examples: https://bit.ly/Htype-gist
}
