% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Htypes.R
\name{is_Hdesc}
\alias{is_Hdesc}
\alias{is_single_Hdesc}
\title{Checks for describe objects}
\usage{
is_Hdesc(x)

is_single_Hdesc(x)
}
\arguments{
\item{x}{an object to test if it is of class `describe`.}
}
\value{
(lgl) is `x` (a single element or a general) `describe`
  object?
}
\description{
These two function are useful to test if an object is of class
[Hmisc][Hmisc::describe].
}
\details{
In `Hmisc` both "single" `describe` objects and lists
of them are of class `describe`. In particular, even if
`Hmisc::describe()` results in a single variable description, it is
 directly the "single" `describe` object and not a list of them with
 only a single `describe` object included!

`is_Hdesc()` test for general inheritance.

`is_single_Hdesc()` test for single instance of a
  `describe` object.
}
\examples{
\dontrun{
  library(Hmisc)
  desc <- describe(mtcars)

  is_Hdesc(desc) # TRUE
  is_Hdesc(desc[[1]]) # TRUE
}
\dontrun{
  is_single_Hdesc(desc) # FALSE
  is_single_Hdesc(desc[[1]]) # TRUE
}
}
\seealso{
[describe][Hmisc::describe]

[is_Hcat], [is_Hcon], [Htype], [Htypes]
}
