% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{please_install}
\alias{please_install}
\title{Please install}
\usage{
please_install(pkgs, install_fun = install.packages, ...)
}
\arguments{
\item{pkgs}{character vector of package(s) to install}

\item{install_fun}{function to use for installing package(s)}

\item{...}{further options for install_fun}
}
\value{
invisible
}
\description{
A polite helper for installing and update packages (quite exactly) taken
from a function used by Hadley Wickham at
`RStudio::conf 2018 - San Diego`.
}
